refactor(lightbox): ♻️ add proper undefined checks
